requirements.txt
fastapi[standard]
psycopg2
alembic
python-dotenv
fastapi_sqlalchemy
googlemaps
pydantic
pytest
starlette